requirements.txt
Django==4.1.7
environs==9.5.0
Pillow==9.4.0
python-telegram-bot==13.*
more-itertools==9.1.0
